+++ src-tauri/src/config/verge.rs
@@ -28,6 +28,9 @@ pub struct IVerge {
     /// `light` or `dark` or `system`
     pub theme_mode: Option<String>,
 
+    /// `red` | `blue` | `green` | `beige`
+    pub theme_palette: Option<String>,
+
     /// tray click event
     pub tray_event: Option<String>,
 
@@ -385,7 +388,8 @@ impl IVerge {
             app_log_max_count: Some(8),
             clash_core: Some("verge-mihomo".into()),
             language: Some(clash_verge_i18n::system_language().into()),
-            theme_mode: Some("system".into()),
+            theme_mode: Some("dark".into()),
+            theme_palette: Some("red".into()),
             #[cfg(not(target_os = "windows"))]
             env_type: Some("bash".into()),
             #[cfg(target_os = "windows")]
@@ -476,6 +480,7 @@ impl IVerge {
 
         patch!(language);
         patch!(theme_mode);
+        patch!(theme_palette);
         patch!(tray_event);
         patch!(env_type);
         patch!(start_page);