+++ src/services/states.ts
@@ -4,6 +4,9 @@ const [ThemeModeProvider, useThemeMode, useSetThemeMode] = createContextState<
   'light' | 'dark'
 >()
 
+const [ThemePaletteProvider, useThemePalette, useSetThemePalette] =
+  createContextState<'red' | 'blue' | 'green' | 'beige'>()
+
 // save the state of each profile item loading
 const [LoadingCacheProvider, useLoadingCache, useSetLoadingCache] =
   createContextState<Set<string>>(new Set())
@@ -16,6 +19,9 @@ export {
   ThemeModeProvider,
   useThemeMode,
   useSetThemeMode,
+  ThemePaletteProvider,
+  useThemePalette,
+  useSetThemePalette,
   LoadingCacheProvider,
   useLoadingCache,
   useSetLoadingCache,